Counts should be usizes imo

+++ src/katajainen.rs
@@ -8,11 +8,11 @@ use std::{mem};
 #[derive(Debug)]
 struct Node {
     weight: usize,
-    leaf_counts: Vec<i32>,
+    leaf_counts: Vec<usize>,
 }
 
 impl Node {
-    fn new(weight: usize, initial_count: i32, capacity: usize) -> Node {
+    fn new(weight: usize, initial_count: usize, capacity: usize) -> Node {
         let mut n = Node {
             weight: weight,
             leaf_counts: Vec::with_capacity(capacity),
@@ -116,7 +116,7 @@ pub fn length_limited_code_lengths(frequencies: &[usize], maxbits: i32) -> Vec<u
     let mut bitlength_value = 1;
     while let Some(leaf_count) = a.next() {
         let next_count = *a.peek().unwrap_or(&0);
-        for leaf in &leaves[next_count as usize..leaf_count as usize] {
+        for leaf in &leaves[next_count..leaf_count] {
             result[leaf.index] = bitlength_value;
         }
         bitlength_value += 1;
